refactor(unlock-app) using a globalwrapper (#8010)
* using a globalwrapper

* moving CONSOLE_MESSAGE

* UnlockApp is now a functional component
